fix counter size

+++ src/cpu.rs
@@ -2,7 +2,7 @@ pub struct CPU {
     pub register_a: u8,
     pub register_x: u8,
     pub status: u8,
-    pub program_counter: u8,
+    pub program_counter: u16,
 }
 
 impl CPU {
